checkout-app/checkout/requirements.txt
flask==3.0.3
requests==2.31.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
